tour/handler/UIEvents: Move UI-handling to handler #653
It's entirely self contained, so no reason not to. More importantly, it
gives us somewhere to hang the documentation examples.
